Update ReadBasicAuth interface signature
Even though `ReadBasicAuth` is not used yet in the code, it looks like `BasicAuthCredentials` is supposed to satisfy it with the `Read()` method. However, the signatures differ in that the interface returns the `error` first instead of `*BasicAuthCredentials`. This could cause bugs in the future when using the interface and also is not considered idiomatic Go code.

Signed-off-by: Michael Gasch <mgasch@vmware.com>