Compare basic-auth with constant time
A third-party security audit of Kubernetes found this issue
in the way they validated basic auth credentials. This change
implements the same check for the basic-auth plugin.

https://github.com/kubernetes/kubernetes/pull/81152

Other plugins are also available such as OAuth2 / OIDC for
the OpenFaaS gateway.

The WWW-Authenticate realm was being set even when auth passed,
this was unnecessary and has been configured to only be set
when auth is required or the header is missing.

Tested with additional unit tests.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>