Update omitempty settings for functions
The deployment JSON was too verbose, where it was setting
null or an empty array for nillable types. This isn't required
and lead to an error when testing openfaas-cloud with the
operator, which didn't tolerate "null" being passed as an
environment variable.

Tested with a unit test to confirm minimum JSON output.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>