Fix issue (#56)
When no valid endpoint is available for a function, a status
of 503 is now sent back to the caller. This endpoint is
considered private and should only be called from the
gateway, after it has already determined the existence of
a function deployment.

Therefore, rather than a 404 error, which was confusing for
users, a 503 "unavailable" is now being returned to the
gateway, so that a clearer understanding of what happened
can be built-up when supporting users.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>