Instrument provider endpoints
Instruments HTTP status code and duration, with additional
path and method parameters. The function name is stripped
from any labels to limit the size of the series.

Tested end to end with port-forwarding from the faas-netes
HTTP port directly from the gateway container.

HttpWriteInterceptor moves here from the gateway, and can
be deleted downstream, to keep functionality in one place.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>