fix: ensure WriteInterceptor always  sets a valid status code
Ensure that the metrics record the correct status code, instead of the
invalid 0 value. This replicates the stdlib behavior of setting the 200
code on first write, if it is the first write.

The stats code is also moved to a private method to ensure that users
always use the Status() method instead.

Resolves #65

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>